refactor app.js
