+++ src/App.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+import {AppNavigation} from './AppNavigation/AppNavigation';
+
+export default class App extends React.Component {
+  render() {
+    return <AppNavigation />;
+  }
+}
